feat: Add SerdeObject support for pasta curves
With the current serialization primitives we have here, pasta curves did
not have any way to be serialized as we did not implement `SerdeObject`
for it.

This PR adds this support which resolves
https://github.com/privacy-scaling-explorations/halo2curves/issues/23.
The code needs to have `unsafe` chunks as we need to transmute memory
in order to be able to access the private fields of the pasta-curves
crate structs.

This should be revisited and re-thought once
https://github.com/privacy-scaling-explorations/halo2/issues/176 is
addressed.